Move frontend connection-related code to src/store/connection
