+++ frontend/src/modules/app/App.tsx
@@ -12,7 +12,7 @@ import { connect } from 'react-redux';
 import { createStructuredSelector } from 'reselect';
 import { initialize } from '../../store/app/actions';
 import { getLocale, getMessages } from '../../store/app/selectors';
-import { FrontendDisplaySetting, ThemeVariant } from '../../store/controller/proto/frontend_pb';
+import { FrontendDisplaySetting, ThemeVariant } from '../../store/proto/frontend_pb';
 import { getFrontendDisplaySetting } from '../../store/controller/selectors';
 import { StoreState } from '../../store/types';
 import { darkTheme, lightTheme } from '../../styles/customTheme';
