+++ frontend/src/modules/dashboard/components/ValueInfo.tsx
@@ -14,7 +14,7 @@ import { getAlarmLimitsRequest } from '../../../store/controller/selectors';
 import { setMultiPopupOpen } from '../../app/Service';
 import { AlarmModal } from '../../controllers';
 import { SelectorType, ValueSelectorDisplay } from '../../displays/ValueSelectorDisplay';
-import { Range } from '../../../store/controller/proto/mcu_pb';
+import { Range } from '../../../store/proto/mcu_pb';
 
 const useStyles = makeStyles((theme: Theme) => ({
   rootParent: {
