+++ frontend/src/modules/displays/ModeBanner.tsx
@@ -6,7 +6,7 @@ import React from 'react';
 import { Grid, Typography, makeStyles, Theme } from '@material-ui/core';
 import { useSelector } from 'react-redux';
 import { getParametersRequestMode } from '../../store/controller/selectors';
-import { VentilationMode } from '../../store/controller/proto/mcu_pb';
+import { VentilationMode } from '../../store/proto/mcu_pb';
 
 const useStyles = makeStyles((theme: Theme) => ({
   normal: {
