+++ frontend/src/modules/displays/MultiStepWizard.tsx
@@ -23,7 +23,7 @@ import { a11yProps, TabPanel } from '../controllers/TabPanel';
 import ValueInfo from '../dashboard/components/ValueInfo';
 import { BPM, LMIN, PERCENT } from '../info/units';
 import { AlarmModal } from '../controllers';
-import { ParametersRequest, AlarmLimitsRequest } from '../../store/controller/proto/mcu_pb';
+import { ParametersRequest, AlarmLimitsRequest } from '../../store/proto/mcu_pb';
 import { MessageType } from '../../store/controller/types';
 import { commitRequest, commitDraftRequest } from '../../store/controller/actions';
 import store from '../../store';
