+++ frontend/src/modules/quickstart/QuickStartPage.tsx
@@ -8,7 +8,7 @@ import { makeStyles, Theme, useTheme } from '@material-ui/core/styles';
 import React, { RefObject, useEffect, useRef } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { commitDraftRequest } from '../../store/controller/actions';
-import { VentilationMode, ParametersRequest } from '../../store/controller/proto/mcu_pb';
+import { VentilationMode, ParametersRequest } from '../../store/proto/mcu_pb';
 import {
   getParametersRequestMode,
   getParametersRequestDraftFiO2,
