+++ frontend/src/store/connection/protocols/application/events.ts
