+++ frontend/src/store/connection/protocols/backend/transport.ts
@@ -1,16 +1,16 @@
-import { MessageType, PBMessage, MessageClass, MessageTypes } from '../../types';
-import {
-  MessageSerializer,
-  makeMessageSerializer,
-  makeMessageDeserializer,
-} from '../transport/messages';
 import {
   ParametersRequest,
   AlarmLimitsRequest,
   ExpectedLogEvent,
   AlarmMuteRequest,
-} from '../../proto/mcu_pb';
-import { FrontendDisplaySetting } from '../../proto/frontend_pb';
+} from '../../../proto/mcu_pb';
+import { FrontendDisplaySetting } from '../../../proto/frontend_pb';
+import { MessageType, PBMessage, MessageClass, MessageTypes } from '../../../controller/types';
+import {
+  MessageSerializer,
+  makeMessageSerializer,
+  makeMessageDeserializer,
+} from '../transport/messages';
 
 // Serialization
