+++ frontend/src/store/controller/types.ts
@@ -12,13 +12,13 @@ import {
   AlarmMuteRequest,
   MCUPowerStatus,
   ScreenStatus,
-} from './proto/mcu_pb';
+} from '../proto/mcu_pb';
 import {
   BackendConnections,
   RotaryEncoder,
   SystemSettingRequest,
   FrontendDisplaySetting,
-} from './proto/frontend_pb';
+} from '../proto/frontend_pb';
 
 // MESSAGES
